+++ example/lib/example.screen.dart
@@ -3,34 +3,47 @@ import 'package:in_app_review/in_app_review.dart';
 import 'package:rating_dialog/rating_dialog.dart';
 
 class _ExampleScreenState extends State<ExampleScreen> {
-  // actual store listing review & rating
-  void _rateAndReviewApp() async {
-    final _inAppReview = InAppReview.instance;
-
-    if (await _inAppReview.isAvailable()) {
-      print('request actual review from store');
-      _inAppReview.requestReview();
-    } else {
-      print('open actual store listing');
-      // TODO: use your own store ids
-      _inAppReview.openStoreListing(
-        appStoreId: '<your app store id>',
-        microsoftStoreId: '<your microsoft store id>',
-      );
-    }
-  }
-
   // show the rating dialog
   void _showRatingDialog() {
+    // actual store listing review & rating
+    void _rateAndReviewApp() async {
+      // refer to: https://pub.dev/packages/in_app_review
+      final _inAppReview = InAppReview.instance;
+
+      if (await _inAppReview.isAvailable()) {
+        print('request actual review from store');
+        _inAppReview.requestReview();
+      } else {
+        print('open actual store listing');
+        // TODO: use your own store ids
+        _inAppReview.openStoreListing(
+          appStoreId: '<your app store id>',
+          microsoftStoreId: '<your microsoft store id>',
+        );
+      }
+    }
+
     final _dialog = RatingDialog(
+      initialRating: 1.0,
       // your app's name?
-      title: 'Rating Dialog',
+      title: Text(
+        'Rating Dialog',
+        textAlign: TextAlign.center,
+        style: const TextStyle(
+          fontSize: 25,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
       // encourage your user to leave a high rating?
-      message:
-          'Tap a star to set your rating. Add more description here if you want.',
+      message: Text(
+        'Tap a star to set your rating. Add more description here if you want.',
+        textAlign: TextAlign.center,
+        style: const TextStyle(fontSize: 15),
+      ),
       // your app's logo?
       image: const FlutterLogo(size: 100),
-      submitButton: 'Submit',
+      submitButtonText: 'Submit',
+      commentHint: 'Set your custom comment hint',
       onCancelled: () => print('cancelled'),
       onSubmitted: (response) {
         print('rating: ${response.rating}, comment: ${response.comment}');
@@ -70,7 +83,7 @@ class _ExampleScreenState extends State<ExampleScreen> {
 }
 
 class ExampleScreen extends StatefulWidget {
-  ExampleScreen();
+  const ExampleScreen();
 
   @override
   _ExampleScreenState createState() => new _ExampleScreenState();